requirements.txt
requests==2.28.2
fastapi==0.89.1
gradio==3.35.2
python-dotenv==1.0.0
openai==0.27.8
uvicorn==0.20.0
google-auth==2.16.0
sendgrid==6.10.0

